Split/extend the magwell classes by size
