+++ addons/jam/CfgMagazineWells.hpp
@@ -1,170 +1,27 @@
 class CfgMagazineWells {
-    class CBA_545x39_AK {
-        BI_mags[] = {
-            "30Rnd_545x39_Mag_F",
-            "30Rnd_545x39_Mag_Green_F",
-            "30Rnd_545x39_Mag_Tracer_F",
-            "30Rnd_545x39_Mag_Tracer_Green_F"
-        };
-    };
-
-    class CBA_556x45_MINIMI {
-        BI_boxes[] = {
-            "200Rnd_556x45_Box_F",
-            "200Rnd_556x45_Box_Red_F",
-            "200Rnd_556x45_Box_Tracer_F",
-            "200Rnd_556x45_Box_Tracer_Red_F"
-        };
-    };
-    class CBA_556x45_STANAG {
-        BI_mags[] = {
-            "30Rnd_556x45_Stanag",
-            "30Rnd_556x45_Stanag_green",
-            "30Rnd_556x45_Stanag_red",
-            "30Rnd_556x45_Stanag_Tracer_Red",
-            "30Rnd_556x45_Stanag_Tracer_Green",
-            "30Rnd_556x45_Stanag_Tracer_Yellow"
-        };
-        BI_drums[] = {
-            "150Rnd_556x45_Drum_Mag_F",
-            "150Rnd_556x45_Drum_Mag_Tracer_F"
-        };
-    };
-
-    class CBA_556x45_STEYR {};        // AUG
-    class CBA_556x45_HK33 {};         // H&K 33/53/93
-    class CBA_556x45_G36 {};
-    class CBA_556x45_SG550 {};
-
-    class CBA_580x42_TYPE95 {
-        BI_mags[] = {
-            "30Rnd_580x42_Mag_F",
-            "30Rnd_580x42_Mag_Tracer_F"
-        };
-        BI_drums[] = {
-            "100Rnd_580x42_Mag_F",
-            "100Rnd_580x42_Mag_Tracer_F"
-        };
-    };
-
-    class CBA_65x39_MX {
-        BI_mags[] = {
-            "30Rnd_65x39_caseless_mag",
-            "30Rnd_65x39_caseless_green",
-            "30Rnd_65x39_caseless_mag_Tracer",
-            "30Rnd_65x39_caseless_green_mag_Tracer",
-            "100Rnd_65x39_caseless_mag",
-            "100Rnd_65x39_caseless_mag_Tracer"
-        };
-    };
+    #include "magwells_556x45.hpp"
+    #include "magwells_545x39.hpp"
+    #include "magwells_580x42.hpp"
+    #include "magwells_65x39.hpp"
+    #include "magwells_762x39.hpp"
+    #include "magwells_762x51.hpp"
+    #include "magwells_762x54.hpp"
+    #include "magwells_9x19.hpp"
+    #include "magwells_357SIG.hpp"
+    #include "magwells_40SW.hpp"
 
     class CBA_68x43_ACR {};         // 6.8mm Remington SPC
     class CBA_75x55_STGW57 {};      // SIG SG 510-1, Stgw. 57
     class CBA_762x35_STANAG {};     // .300 Blackout
-
-    class CBA_762x39_AK {
-        BI_mags[] = {
-            "30Rnd_762x39_Mag_F",
-            "30Rnd_762x39_Mag_Green_F",
-            "30Rnd_762x39_Mag_Tracer_F",
-            "30Rnd_762x39_Mag_Tracer_Green_F"
-        };
-    };
-
-    class CBA_762x51_M14 {
-        BI_mags[] = {
-            "20Rnd_762x51_Mag"
-        };
-    };
-    class CBA_762x51_HK417 {
-        BI_mags[] = {
-            "20Rnd_762x51_Mag"
-        };
-    };
-    class CBA_762x51_SR25 {};       // M110, SIG716
-    class CBA_762x51_SCAR {};
-    class CBA_762x51_FAL {};
-    class CBA_762x51_G3 {};
-    class CBA_762x51_SIGAMT {};    // SIG 510-4, AMT
-
-    class CBA_762x51_LINKS {
-        BI_belts[] = {
-            "150Rnd_762x51_Box",
-            "150Rnd_762x51_Box_Tracer"
-        };
-    };
-
-    class CBA_762x54R_SVD {
-        BI_mags[] = {
-            "10Rnd_762x54_Mag"
-        };
-    };
-    class CBA_762x54R_LINKS {
-        BI_belts[] = {
-            "150Rnd_762x54_Box",
-            "150Rnd_762x54_Box_Tracer"
-        };
-    };
-
     class CBA_792x57_LINKS {};      // MG42
     class CBA_300WM_AI {};          // AI .300 Winchester Magnum
     class CBA_338LM_AI {};          // AI .338 Lapua Magnum
     class CBA_50BMG_M107 {};        // M82, M107, G82
-
-    class CBA_9x19_MP5 {};          // H&K MP5
-    class CBA_9x19_STEYR {};        // AUG SMG, MPi, TMP
-    class CBA_9x19_P226 {};         // SIG P226
-    class CBA_9x19_P228 {};         // SIG P228
-    class CBA_9x19_P239 {};         // SIG P239
-
     class CBA_9x39_VSS {};          // Vintorez, Val
-
-    class CBA_357SIG_P226 {};       // SIG P226
-    class CBA_357SIG_P229 {};       // SIG P229
-    class CBA_357SIG_P239 {};       // SIG P239
-
-    class CBA_40SW_P226 {};         // SIG P226 .40 S&W
-    class CBA_40SW_P229 {};         // SIG P229 .40 S&W
-    class CBA_40SW_P239 {};         // SIG P239 .40 S&W
-
     class CBA_10x25_MP5 {};         // H&K MP5/10 10mm Auto
     class CBA_12g_SAIGA {};
 
-    class CBA_40mm_M203 {
-        BI_grenades[] = {
-            "1Rnd_HE_Grenade_shell",
-            "1Rnd_Smoke_Grenade_shell",
-            "1Rnd_SmokeRed_Grenade_shell",
-            "1Rnd_SmokeGreen_Grenade_shell",
-            "1Rnd_SmokeYellow_Grenade_shell",
-            "1Rnd_SmokePurple_Grenade_shell",
-            "1Rnd_SmokeBlue_Grenade_shell",
-            "1Rnd_SmokeOrange_Grenade_shell",
-            "UGL_FlareWhite_F",
-            "UGL_FlareGreen_F",
-            "UGL_FlareRed_F",
-            "UGL_FlareYellow_F",
-            "UGL_FlareCIR_F"
-        };
-    };
-    class CBA_40mm_3GL {
-        BI_grenades[] = {
-            "3Rnd_HE_Grenade_shell",
-            "3Rnd_Smoke_Grenade_shell",
-            "3Rnd_SmokeRed_Grenade_shell",
-            "3Rnd_SmokeGreen_Grenade_shell",
-            "3Rnd_SmokeYellow_Grenade_shell",
-            "3Rnd_SmokePurple_Grenade_shell",
-            "3Rnd_SmokeBlue_Grenade_shell",
-            "3Rnd_SmokeOrange_Grenade_shell",
-            "3Rnd_UGL_FlareWhite_F",
-            "3Rnd_UGL_FlareGreen_F",
-            "3Rnd_UGL_FlareRed_F",
-            "3Rnd_UGL_FlareYellow_F",
-            "3Rnd_UGL_FlareCIR_F"
-        };
-    };
-    class CBA_40mm_GP {};
+    #include "magwells_40mm.hpp"
 
     class CBA_RPG7 {
         BI_rockets[] = {
