+++ addons/jam/CfgWeapons.hpp
@@ -5,29 +5,29 @@ class CfgWeapons {
     class Launcher_Base_F;
 
     class mk20_base_F: Rifle_Base_F {
-        magazineWell[] = {"CBA_556x45_STANAG"};
+        magazineWell[] = {"CBA_556x45_STANAG", "CBA_556x45_STANAG_L", "CBA_556x45_STANAG_XL", "CBA_556x45_STANAG_2D"};
     };
     class SDAR_base_F: Rifle_Base_F {
         magazineWell[] = {"CBA_556x45_STANAG"};
     };
     class Tavor_base_F: Rifle_Base_F {
-        magazineWell[] = {"CBA_556x45_STANAG"};
+        magazineWell[] = {"CBA_556x45_STANAG", "CBA_556x45_STANAG_L", "CBA_556x45_STANAG_XL", "CBA_556x45_STANAG_2D"};
     };
     class arifle_SPAR_01_base_F : Rifle_Base_F {
-        magazineWell[] = {"CBA_556x45_STANAG"};
+        magazineWell[] = {"CBA_556x45_STANAG", "CBA_556x45_STANAG_L", "CBA_556x45_STANAG_XL", "CBA_556x45_STANAG_2D", "CBA_556x45_STANAG_2D_XL"};
     };
     class arifle_SPAR_02_base_F : Rifle_Base_F {
-        magazineWell[] = {"CBA_556x45_STANAG"};
+        magazineWell[] = {"CBA_556x45_STANAG", "CBA_556x45_STANAG_L", "CBA_556x45_STANAG_XL", "CBA_556x45_STANAG_2D", "CBA_556x45_STANAG_2D_XL"};
     };
 
     class UGL_F : GrenadeLauncher {
-        magazineWell[] = {"CBA_40mm_M203"};
+        magazineWell[] = {"CBA_40mm_M203", "CBA_40mm_EGLM"};
     };
 
     class arifle_MX_Base_F : Rifle_Base_F {
-        magazineWell[] = {"CBA_65x39_MX"};
+        magazineWell[] = {"CBA_65x39_MX", "CBA_65x39_MX_XL"};
         class GL_3GL_F : UGL_F {
-            magazineWell[] = {"CBA_40mm_3GL", "CBA_40mm_M203"};
+            magazineWell[] = {"CBA_40mm_3GL", "CBA_40mm_M203", "CBA_40mm_EGLM"};
         };
     };
 
@@ -60,7 +60,7 @@ class CfgWeapons {
     };
 
     class arifle_AKS_base_F : Rifle_Base_F {
-        magazineWell[] = {"CBA_545x39_AK"};
+        magazineWell[] = {"CBA_545x39_AK", "CBA_545x39_AK_L"};
     };
 
     class LMG_03_base_F : Rifle_Long_Base_F {
