+++ addons/jam/magwells_357SIG.hpp
@@ -1,0 +1,3 @@
+    class CBA_357SIG_P226 {};       // SIG P226
+    class CBA_357SIG_P229 {};       // SIG P229
+    class CBA_357SIG_P239 {};       // SIG P239
