+++ addons/jam/magwells_40SW.hpp
@@ -1,0 +1,3 @@
+    class CBA_40SW_P226 {};         // SIG P226 .40 S&W
+    class CBA_40SW_P229 {};         // SIG P229 .40 S&W
+    class CBA_40SW_P239 {};         // SIG P239 .40 S&W
