+++ addons/jam/magwells_40mm.hpp
@@ -1,0 +1,38 @@
+    class CBA_40mm_M203 {
+        BI_grenades[] = {
+            "1Rnd_HE_Grenade_shell",
+            "1Rnd_Smoke_Grenade_shell",
+            "1Rnd_SmokeRed_Grenade_shell",
+            "1Rnd_SmokeGreen_Grenade_shell",
+            "1Rnd_SmokeYellow_Grenade_shell",
+            "1Rnd_SmokePurple_Grenade_shell",
+            "1Rnd_SmokeBlue_Grenade_shell",
+            "1Rnd_SmokeOrange_Grenade_shell",
+            "UGL_FlareWhite_F",
+            "UGL_FlareGreen_F",
+            "UGL_FlareRed_F",
+            "UGL_FlareYellow_F",
+            "UGL_FlareCIR_F"
+        };
+    };
+
+    class CBA_40mm_3GL {
+        BI_grenades[] = {
+            "3Rnd_HE_Grenade_shell",
+            "3Rnd_Smoke_Grenade_shell",
+            "3Rnd_SmokeRed_Grenade_shell",
+            "3Rnd_SmokeGreen_Grenade_shell",
+            "3Rnd_SmokeYellow_Grenade_shell",
+            "3Rnd_SmokePurple_Grenade_shell",
+            "3Rnd_SmokeBlue_Grenade_shell",
+            "3Rnd_SmokeOrange_Grenade_shell",
+            "3Rnd_UGL_FlareWhite_F",
+            "3Rnd_UGL_FlareGreen_F",
+            "3Rnd_UGL_FlareRed_F",
+            "3Rnd_UGL_FlareYellow_F",
+            "3Rnd_UGL_FlareCIR_F"
+        };
+    };
+
+    class CBA_40mm_EGLM {};     // for longer grenades that can only fit side breech-loading
+    class CBA_40mm_GP {};       // GP-25, GP-30
