+++ addons/jam/magwells_545x39.hpp
@@ -1,0 +1,10 @@
+    class CBA_545x39_AK {
+        BI_mags[] = {
+            "30Rnd_545x39_Mag_F",
+            "30Rnd_545x39_Mag_Green_F",
+            "30Rnd_545x39_Mag_Tracer_F",
+            "30Rnd_545x39_Mag_Tracer_Green_F"
+        };
+    };
+
+    class CBA_545x39_AK_L {};    // 45rnd RPK-74
