+++ addons/jam/magwells_556x45.hpp
@@ -1,0 +1,35 @@
+    class CBA_556x45_MINIMI {
+        BI_boxes[] = {
+            "200Rnd_556x45_Box_F",
+            "200Rnd_556x45_Box_Red_F",
+            "200Rnd_556x45_Box_Tracer_F",
+            "200Rnd_556x45_Box_Tracer_Red_F"
+        };
+    };
+
+    class CBA_556x45_STANAG {
+        BI_mags[] = {
+            "30Rnd_556x45_Stanag",
+            "30Rnd_556x45_Stanag_green",
+            "30Rnd_556x45_Stanag_red",
+            "30Rnd_556x45_Stanag_Tracer_Red",
+            "30Rnd_556x45_Stanag_Tracer_Green",
+            "30Rnd_556x45_Stanag_Tracer_Yellow"
+        };
+    };
+
+    class CBA_556x45_STANAG_L {};           // Magpul PMAG 40, Surefire MAG5-60
+    class CBA_556x45_STANAG_XL {};          // Surefire MAG5-100
+    class CBA_556x45_STANAG_2D {};          // twin-drum 100rnd Beta C-MAG
+
+    class CBA_556x45_STANAG_2D_XL {         // twin-drum 150rnd Armatac SAW-MAG
+        BI_mags[] = {
+            "150Rnd_556x45_Drum_Mag_F",
+            "150Rnd_556x45_Drum_Mag_Tracer_F"
+        };
+    };
+
+    class CBA_556x45_STEYR {};        // AUG
+    class CBA_556x45_HK33 {};         // H&K 33/53/93
+    class CBA_556x45_G36 {};
+    class CBA_556x45_SG550 {};
