+++ addons/jam/magwells_580x42.hpp
@@ -1,0 +1,10 @@
+    class CBA_580x42_TYPE95 {
+        BI_mags[] = {
+            "30Rnd_580x42_Mag_F",
+            "30Rnd_580x42_Mag_Tracer_F"
+        };
+        BI_drums[] = {
+            "100Rnd_580x42_Mag_F",
+            "100Rnd_580x42_Mag_Tracer_F"
+        };
+    };
