+++ addons/jam/magwells_65x39.hpp
@@ -1,0 +1,14 @@
+    class CBA_65x39_MX {
+        BI_mags[] = {
+            "30Rnd_65x39_caseless_mag",
+            "30Rnd_65x39_caseless_green",
+            "30Rnd_65x39_caseless_mag_Tracer",
+            "30Rnd_65x39_caseless_green_mag_Tracer"
+        };
+    };
+    class CBA_65x39_MX_XL {
+        BI_mags[] = {
+            "100Rnd_65x39_caseless_mag",
+            "100Rnd_65x39_caseless_mag_Tracer"
+        };
+    };
