+++ addons/jam/magwells_762x39.hpp
@@ -1,0 +1,8 @@
+    class CBA_762x39_AK {
+        BI_mags[] = {
+            "30Rnd_762x39_Mag_F",
+            "30Rnd_762x39_Mag_Green_F",
+            "30Rnd_762x39_Mag_Tracer_F",
+            "30Rnd_762x39_Mag_Tracer_Green_F"
+        };
+    };
