+++ addons/jam/magwells_762x51.hpp
@@ -1,0 +1,24 @@
+    class CBA_762x51_M14 {
+        BI_mags[] = {
+            "20Rnd_762x51_Mag"
+        };
+    };
+
+    class CBA_762x51_HK417 {
+        BI_mags[] = {
+            "20Rnd_762x51_Mag"
+        };
+    };
+
+    class CBA_762x51_SR25 {};       // M110, SIG716
+    class CBA_762x51_SCAR {};
+    class CBA_762x51_FAL {};
+    class CBA_762x51_G3 {};
+    class CBA_762x51_SIGAMT {};    // SIG 510-4, AMT
+
+    class CBA_762x51_LINKS {
+        BI_belts[] = {
+            "150Rnd_762x51_Box",
+            "150Rnd_762x51_Box_Tracer"
+        };
+    };
