+++ addons/jam/magwells_762x54.hpp
@@ -1,0 +1,12 @@
+    class CBA_762x54R_SVD {
+        BI_mags[] = {
+            "10Rnd_762x54_Mag"
+        };
+    };
+
+    class CBA_762x54R_LINKS {
+        BI_belts[] = {
+            "150Rnd_762x54_Box",
+            "150Rnd_762x54_Box_Tracer"
+        };
+    };
