+++ addons/jam/magwells_9x19.hpp
@@ -1,0 +1,5 @@
+    class CBA_9x19_MP5 {};          // H&K MP5
+    class CBA_9x19_STEYR {};        // AUG SMG, MPi, TMP
+    class CBA_9x19_P226 {};         // SIG P226
+    class CBA_9x19_P228 {};         // SIG P228
+    class CBA_9x19_P239 {};         // SIG P239
